relevant logic extracted into ColumnWriter
