+++ src/Parquet.Runner/Parquet.Runner.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="LogMagic" Version="2.4.29" />
+    <PackageReference Include="LogMagic" Version="2.5.61" />
   </ItemGroup>
 
   <ItemGroup>
