+++ src/Parquet/ParquetWriter.cs
@@ -1,33 +1,8 @@
-﻿/* MIT License
- *
- * Copyright (c) 2017 Elastacloud Limited
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-using System;
+﻿using System;
 using System.IO;
 using Parquet.File;
 using System.Collections.Generic;
 using System.Linq;
-using Parquet.File.Values;
-using Parquet.File.Data;
 using System.Collections;
 using Parquet.Data;
 
@@ -42,18 +17,8 @@ namespace Parquet
       private readonly FileMetadataBuilder _meta;
       private readonly ParquetOptions _formatOptions;
       private readonly WriterOptions _writerOptions;
-      private readonly SchemaElement _definitionsSchema = new SchemaElement<bool>("definitions");
-      private readonly IValuesWriter _plainWriter;
-      private readonly IValuesWriter _rleWriter;
-      private readonly IValuesWriter _dicWriter;
       private Schema _existingSchema;
       private bool _dataWritten;
-
-      private struct PageTag
-      {
-         public int HeaderSize;
-         public Thrift.PageHeader HeaderMeta;
-      }
 
       /// <summary>
       /// Creates an instance of parquet writer on top of a stream
@@ -70,10 +35,6 @@ namespace Parquet
          _formatOptions = formatOptions ?? new ParquetOptions();
          _writerOptions = writerOptions ?? new WriterOptions();
          _meta = new FileMetadataBuilder();
-
-         _plainWriter = new PlainValuesWriter(_formatOptions);
-         _rleWriter = new RunLengthBitPackingHybridValuesWriter();
-         _dicWriter = new PlainDictionaryValuesWriter(_rleWriter);
       }
 
       /// <summary>
@@ -95,10 +56,21 @@ namespace Parquet
             count = Math.Min(_writerOptions.RowGroupsSize, dataSet.Count - offset);
             Thrift.RowGroup rg = _meta.AddRowGroup();
             long rgStartPos = _output.Position;
-            rg.Columns = dataSet.Schema.Elements
+
+            rg.Columns = new List<Thrift.ColumnChunk>();
+            foreach(SchemaElement se in dataSet.Schema.Flatten())
+            {
+               var cw = new ColumnWriter(_output, ThriftStream, _meta, se, compression, _formatOptions, _writerOptions);
+               IList values = dataSet.GetColumn(se.Name, offset, count);
+               ColumnStats cs = stats.GetColumnStats(se);
+               Thrift.ColumnChunk chunk = cw.Write(offset, count, values, cs);
+               rg.Columns.Add(chunk);
+            }
+
+            /*rg.Columns = dataSet.Schema.Elements
                .Select(c =>
                   Write(c, dataSet.GetColumn(c.Name, offset, count), compression, stats.GetColumnStats(c)))
-               .ToList();
+               .ToList();*/
 
             //row group's size is a sum of _uncompressed_ sizes of all columns in it, including the headers
             //luckily ColumnChunk already contains sizes of page+header in it's meta
@@ -188,165 +160,6 @@ namespace Parquet
                writer.Write(dataSet, compression);
             }
          }
-      }
-
-      private Thrift.ColumnChunk Write(SchemaElement schema, IList values,
-         CompressionMethod compression,
-         ColumnStats stats)
-      {
-         Thrift.ColumnChunk chunk = _meta.AddColumnChunk(compression, _output, schema, values.Count);
-         Thrift.PageHeader ph = _meta.CreateDataPage(values.Count);
-
-         List<PageTag> pages = WriteValues(schema, values, ph, compression, stats);
-
-         //the following counters must include both data size and header size
-         chunk.Meta_data.Total_compressed_size = pages.Sum(p => p.HeaderMeta.Compressed_page_size + p.HeaderSize);
-         chunk.Meta_data.Total_uncompressed_size = pages.Sum(p => p.HeaderMeta.Uncompressed_page_size + p.HeaderSize);
-
-         return chunk;
-      }
-
-      private List<PageTag> WriteValues(SchemaElement schema, IList values, Thrift.PageHeader ph, CompressionMethod compression, ColumnStats stats)
-      {
-         var result = new List<PageTag>();
-         byte[] dictionaryPageBytes = null;
-         int dictionaryPageCount = 0;
-         byte[] dataPageBytes;
-
-         //flatten values if the field is repeatable
-         if (schema.IsRepeated)
-         {
-            values = FlattenRepeatables(values, schema);
-         }
-
-         using (var ms = new MemoryStream())
-         {
-            using (var writer = new BinaryWriter(ms))
-            {
-               //write repetitions
-               if (schema.IsRepeated)
-               {
-                  List<int> repetitions = CreateRepetitions(values, schema);
-                  _rleWriter.Write(writer, _definitionsSchema, repetitions, out IList nullExtra);
-               }
-
-               //write definitions
-               if (schema.HasNulls || schema.IsRepeated)
-               {
-                  CreateDefinitions(values, schema, out IList newValues, out List<int> definitions);
-                  values = newValues;
-
-                  _rleWriter.Write(writer, _definitionsSchema, definitions, out IList nullExtra);
-               }
-
-               //write data
-               if (!_writerOptions.UseDictionaryEncoding || !_dicWriter.Write(writer, schema, values, out IList dicValues))
-               {
-                  _plainWriter.Write(writer, schema, values, out IList plainExtra);
-               }
-               else
-               {
-                  dictionaryPageCount = dicValues.Count;
-                  ph.Data_page_header.Encoding = Thrift.Encoding.PLAIN_DICTIONARY;
-                  using (var dms = new MemoryStream())
-                     using(var dwriter = new BinaryWriter(dms))
-                  {
-                     _plainWriter.Write(dwriter, schema, dicValues, out IList t0);
-                     dictionaryPageBytes = dms.ToArray();
-                  }
-               }
-
-               dataPageBytes = ms.ToArray();
-            }
-         }
-
-         if(dictionaryPageBytes != null)
-         {
-            Thrift.PageHeader dph = _meta.CreateDictionaryPage(dictionaryPageCount);
-            dictionaryPageBytes = Compress(dph, dictionaryPageBytes, compression);
-            int dictionaryHeaderSize = Write(dph, dictionaryPageBytes);
-            result.Add(new PageTag { HeaderSize = dictionaryHeaderSize, HeaderMeta = dph });
-         }
-
-         dataPageBytes = Compress(ph, dataPageBytes, compression);
-         int dataHeaderSize = Write(ph, dataPageBytes);
-         result.Add(new PageTag { HeaderSize = dataHeaderSize, HeaderMeta = ph });
-
-         return result;
-      }
-
-      private void CreateDefinitions(IList values, SchemaElement schema, out IList nonNullableValues, out List<int> definitions)
-      {
-         nonNullableValues = TypeFactory.Create(schema, _formatOptions, false);
-         definitions = new List<int>();
-
-         foreach(object value in values)
-         {
-            if(value == null)
-            {
-               definitions.Add(0);
-            }
-            else
-            {
-               definitions.Add(1);
-               nonNullableValues.Add(value);
-            }
-         }
-      }
-
-      private List<int> CreateRepetitions(IList values, SchemaElement schema)
-      {
-         var result = new List<int>();
-
-         result.Add(0);
-         result.AddRange(Enumerable.Repeat(1, values.Count - 1));
-
-         return result;
-      }
-
-      private IList FlattenRepeatables(IList values, SchemaElement se)
-      {
-         IList result = TypeFactory.Create(se.ElementType, se.IsNullable);
-
-         foreach(IEnumerable elementValues in values)
-         {
-            result.AddRange(elementValues);
-         }
-
-         return result;
-      }
-
-      private int Write(Thrift.PageHeader ph, byte[] data)
-      {
-         int headerSize = ThriftStream.Write(ph);
-         _output.Write(data, 0, data.Length);
-         return headerSize;
-      }
-
-      private byte[] Compress(Thrift.PageHeader ph, byte[] data, CompressionMethod compression)
-      {
-         //note that page size numbers do not include header size by spec
-
-         ph.Uncompressed_page_size = data.Length;
-         byte[] result;
-
-         if (compression != CompressionMethod.None)
-         {
-            IDataWriter writer = DataFactory.GetWriter(compression);
-            using (var ms = new MemoryStream())
-            {
-               writer.Write(data, ms);
-               result = ms.ToArray();
-            }
-            ph.Compressed_page_size = result.Length;
-         }
-         else
-         {
-            ph.Compressed_page_size = ph.Uncompressed_page_size;
-            result = data;
-         }
-
-         return result;
       }
 
       private void WriteMagic()
